--- a/asm/dscal_amd64.go
+++ b/asm/dscal_amd64.go
@@ -9,3 +9,7 @@ package asm
 func DscalUnitary(alpha float64, x []float64)
 
 func DscalUnitaryTo(dst []float64, alpha float64, x []float64)
+
+func DscalInc(alpha float64, x []float64, n, incX, ix uintptr)
+
+func DscalIncTo(dst []float64, incDst, idst uintptr, alpha float64, x []float64, n, incX, ix uintptr)