--- a/asm/sscal.go
+++ b/asm/sscal.go
@@ -17,3 +17,18 @@ func SscalUnitaryTo(dst []float32, alpha float32, x []float32) {
 		dst[i] = alpha * v
 	}
 }
+
+func SscalInc(alpha float32, x []float32, n, incX, ix uintptr) {
+	for i := 0; i < int(n); i++ {
+		x[ix] *= alpha
+		ix += incX
+	}
+}
+
+func SscalIncTo(dst []float32, incDst, idst uintptr, alpha float32, x []float32, n, incX, ix uintptr) {
+	for i := 0; i < int(n); i++ {
+		dst[idst] = alpha * x[ix]
+		ix += incX
+		idst += incDst
+	}
+}