… asm/dscalinc_amd64.s
… asm/dscalincto_amd64.s