I also added benchmarks for all the four `Dscal` routines and the results are ... errr ... ehm ... well ... below.

The `To` versions behave most reasonably and as expected, and the benefit of assembly is clear. The non-`To` versions are unpredictable and in general much slower than corresponding `To` versions. The `ToX` benchmark suggests that all this mess comes from memory and cache misses (I'm just guessing, though). Fixing this and getting a predictable and consistent performance improvement is beyond my skills. However, at least floats and mat64 could benefit from the `To` versions. I'm curious to read your opinions, @kortschak @btracey .

I haven't checked thoroughly but I think this strangeness does not happen with Dapxy.

```
$ benchcmp dscal_noasm.txt dscal.txt
benchmark                              old ns/op     new ns/op     delta
BenchmarkDscalUnitaryN1-6              45.6          44.9          -1.54%
BenchmarkDscalUnitaryN2-6              85.5          84.1          -1.64%
BenchmarkDscalUnitaryN3-6              125           124           -0.80%
BenchmarkDscalUnitaryN4-6              165           83.2          -49.58%
BenchmarkDscalUnitaryN10-6             407           243           -40.29%
BenchmarkDscalUnitaryN100-6            4027          1938          -51.87%
BenchmarkDscalUnitaryN1000-6           18856         14191         -24.74%
BenchmarkDscalUnitaryN10000-6          6927          2415          -65.14%
BenchmarkDscalUnitaryN100000-6         79901         45959         -42.48%
BenchmarkDscalUnitaryToN1-6            3.57          3.42          -4.20%
BenchmarkDscalUnitaryToN2-6            4.48          3.84          -14.29%
BenchmarkDscalUnitaryToN3-6            5.24          4.51          -13.93%
BenchmarkDscalUnitaryToN4-6            5.99          3.71          -38.06%
BenchmarkDscalUnitaryToN10-6           11.0          6.04          -45.09%
BenchmarkDscalUnitaryToN100-6          82.1          20.5          -75.03%
BenchmarkDscalUnitaryToN1000-6         685           160           -76.64%
BenchmarkDscalUnitaryToN10000-6        6763          2946          -56.44%
BenchmarkDscalUnitaryToN100000-6       78938         51670         -34.54%
BenchmarkDscalUnitaryToXN1-6           3.40          45.4          +1235.29%
BenchmarkDscalUnitaryToXN2-6           85.1          84.6          -0.59%
BenchmarkDscalUnitaryToXN3-6           124           4.20          -96.61%
BenchmarkDscalUnitaryToXN4-6           5.77          82.6          +1331.54%
BenchmarkDscalUnitaryToXN10-6          402           5.76          -98.57%
BenchmarkDscalUnitaryToXN100-6         3995          1928          -51.74%
BenchmarkDscalUnitaryToXN1000-6        40047         160           -99.60%
BenchmarkDscalUnitaryToXN10000-6       6778          155387        +2192.52%
BenchmarkDscalUnitaryToXN100000-6      3542391       613918        -82.67%
BenchmarkDscalIncN1Inc1-6              3.95          46.2          +1069.62%
BenchmarkDscalIncN2Inc1-6              86.7          47.3          -45.44%
BenchmarkDscalIncN2Inc2-6              5.22          4.62          -11.49%
BenchmarkDscalIncN2Inc4-6              86.7          4.63          -94.66%
BenchmarkDscalIncN2Inc10-6             5.23          47.3          +804.40%
BenchmarkDscalIncN3Inc1-6              126           86.6          -31.27%
BenchmarkDscalIncN3Inc2-6              126           86.6          -31.27%
BenchmarkDscalIncN3Inc4-6              125           86.6          -30.72%
BenchmarkDscalIncN3Inc10-6             126           86.6          -31.27%
BenchmarkDscalIncN4Inc1-6              166           6.49          -96.09%
BenchmarkDscalIncN4Inc2-6              168           88.0          -47.62%
BenchmarkDscalIncN4Inc4-6              7.91          6.28          -20.61%
BenchmarkDscalIncN4Inc10-6             166           88.7          -46.57%
BenchmarkDscalIncN10Inc1-6             408           210           -48.53%
BenchmarkDscalIncN10Inc2-6             12.5          10.3          -17.60%
BenchmarkDscalIncN10Inc4-6             12.5          210           +1580.00%
BenchmarkDscalIncN10Inc10-6            12.5          210           +1580.00%
BenchmarkDscalIncN1000Inc1-6           40119         20364         -49.24%
BenchmarkDscalIncN1000Inc2-6           40136         1001          -97.51%
BenchmarkDscalIncN1000Inc4-6           787           1001          +27.19%
BenchmarkDscalIncN1000Inc10-6          40062         1844          -95.40%
BenchmarkDscalIncN100000Inc1-6         79311         102941        +29.79%
BenchmarkDscalIncN100000Inc2-6         111632        861567        +671.79%
BenchmarkDscalIncN100000Inc4-6         177350        158032        -10.89%
BenchmarkDscalIncN100000Inc10-6        341063        304160        -10.82%
BenchmarkDscalIncN100000IncM1-6        3520289       1743826       -50.46%
BenchmarkDscalIncN100000IncM2-6        112741        110403        -2.07%
BenchmarkDscalIncN100000IncM4-6        2909581       787760        -72.93%
BenchmarkDscalIncN100000IncM10-6       2847500       309873        -89.12%
BenchmarkDscalIncToN1Inc1-6            5.16          4.64          -10.08%
BenchmarkDscalIncToN2Inc1-6            6.33          5.57          -12.01%
BenchmarkDscalIncToN2Inc2-6            6.33          5.56          -12.16%
BenchmarkDscalIncToN2Inc4-6            6.33          5.58          -11.85%
BenchmarkDscalIncToN2Inc10-6           6.33          5.56          -12.16%
BenchmarkDscalIncToN3Inc1-6            7.86          5.92          -24.68%
BenchmarkDscalIncToN3Inc2-6            7.79          5.91          -24.13%
BenchmarkDscalIncToN3Inc4-6            7.79          5.92          -24.01%
BenchmarkDscalIncToN3Inc10-6           7.79          5.91          -24.13%
BenchmarkDscalIncToN4Inc1-6            8.80          7.22          -17.95%
BenchmarkDscalIncToN4Inc2-6            8.80          8.15          -7.39%
BenchmarkDscalIncToN4Inc4-6            8.80          7.22          -17.95%
BenchmarkDscalIncToN4Inc10-6           8.80          7.22          -17.95%
BenchmarkDscalIncToN10Inc1-6           13.8          12.0          -13.04%
BenchmarkDscalIncToN10Inc2-6           13.8          12.0          -13.04%
BenchmarkDscalIncToN10Inc4-6           13.8          12.0          -13.04%
BenchmarkDscalIncToN10Inc10-6          13.8          12.0          -13.04%
BenchmarkDscalIncToN1000Inc1-6         870           1004          +15.40%
BenchmarkDscalIncToN1000Inc2-6         871           1006          +15.50%
BenchmarkDscalIncToN1000Inc4-6         1132          1179          +4.15%
BenchmarkDscalIncToN1000Inc10-6        2317          2364          +2.03%
BenchmarkDscalIncToN100000Inc1-6       91345         101280        +10.88%
BenchmarkDscalIncToN100000Inc2-6       124408        115510        -7.15%
BenchmarkDscalIncToN100000Inc4-6       206598        206585        -0.01%
BenchmarkDscalIncToN100000Inc10-6      810037        815643        +0.69%
BenchmarkDscalIncToN100000IncM1-6      92547         100667        +8.77%
BenchmarkDscalIncToN100000IncM2-6      127906        115903        -9.38%
BenchmarkDscalIncToN100000IncM4-6      208331        207308        -0.49%
BenchmarkDscalIncToN100000IncM10-6     824383        788302        -4.38%
```