@kortschak 2016-01-06
Worth asking for advice on golang-nuts?