@vladimir-ch 2016-01-06
Let me run the benchmarks again and I will consider it tomorrow based on the results. golang-nuts is a too big audience for my introvert personality.