@btracey 2016-01-06
Here are the benchmarks on my computer

````
brendan:~/Documents/mygo/src/github.com/gonum$ benchcmp noasm.txt asm.txt
benchmark                              old ns/op     new ns/op     delta         
BenchmarkDscalUnitaryN1-8              53.0          50.3          -5.09%        
BenchmarkDscalUnitaryN2-8              103           100           -2.91%        
BenchmarkDscalUnitaryN3-8              152           147           -3.29%        
BenchmarkDscalUnitaryN4-8              201           97.1          -51.69%       
BenchmarkDscalUnitaryN10-8             492           283           -42.48%       
BenchmarkDscalUnitaryN100-8            4800          2326          -51.54%       
BenchmarkDscalUnitaryN1000-8           47948         17196         -64.14%       
BenchmarkDscalUnitaryN10000-8          236353        2666          -98.87%       
BenchmarkDscalUnitaryN100000-8         108539        43684         -59.75%       
BenchmarkDscalUnitaryToN1-8            4.75          4.38          -7.79%        
BenchmarkDscalUnitaryToN2-8            5.73          5.05          -11.87%       
BenchmarkDscalUnitaryToN3-8            6.65          5.88          -11.58%       
BenchmarkDscalUnitaryToN4-8            7.72          4.94          -36.01%       
BenchmarkDscalUnitaryToN10-8           12.5          7.77          -37.84%       
BenchmarkDscalUnitaryToN100-8          103           27.3          -73.50%       
BenchmarkDscalUnitaryToN1000-8         941           243           -74.18%       
BenchmarkDscalUnitaryToN10000-8        9598          3415          -64.42%       
BenchmarkDscalUnitaryToN100000-8       95698         52449         -45.19%       
BenchmarkDscalUnitaryToXN1-8           4.46          53.1          +1090.58%     
BenchmarkDscalUnitaryToXN2-8           106           102           -3.77%        
BenchmarkDscalUnitaryToXN3-8           5.99          5.94          -0.83%        
BenchmarkDscalUnitaryToXN4-8           197           97.5          -50.51%       
BenchmarkDscalUnitaryToXN10-8          12.5          6.64          -46.88%       
BenchmarkDscalUnitaryToXN100-8         105           2323          +2112.38%     
BenchmarkDscalUnitaryToXN1000-8        47364         240           -99.49%       
BenchmarkDscalUnitaryToXN10000-8       465158        186808        -59.84%       
BenchmarkDscalUnitaryToXN100000-8      4208111       721102        -82.86%       
BenchmarkDscalIncN1Inc1-8              55.0          53.4          -2.91%        
BenchmarkDscalIncN2Inc1-8              107           53.0          -50.47%       
BenchmarkDscalIncN2Inc2-8              6.04          5.20          -13.91%       
BenchmarkDscalIncN2Inc4-8              105           5.08          -95.16%       
BenchmarkDscalIncN2Inc10-8             6.18          53.5          +765.70%      
BenchmarkDscalIncN3Inc1-8              148           101           -31.76%       
BenchmarkDscalIncN3Inc2-8              152           101           -33.55%       
BenchmarkDscalIncN3Inc4-8              154           99.8          -35.19%       
BenchmarkDscalIncN3Inc10-8             153           99.8          -34.77%       
BenchmarkDscalIncN4Inc1-8              201           7.59          -96.22%       
BenchmarkDscalIncN4Inc2-8              202           106           -47.52%       
BenchmarkDscalIncN4Inc4-8              201           7.29          -96.37%       
BenchmarkDscalIncN4Inc10-8             193           104           -46.11%       
BenchmarkDscalIncN10Inc1-8             14.2          242           +1604.23%     
BenchmarkDscalIncN10Inc2-8             14.4          11.3          -21.53%       
BenchmarkDscalIncN10Inc4-8             14.5          243           +1575.86%     
BenchmarkDscalIncN10Inc10-8            477           246           -48.43%       
BenchmarkDscalIncN1000Inc1-8           47414         23582         -50.26%       
BenchmarkDscalIncN1000Inc2-8           47600         1073          -97.75%       
BenchmarkDscalIncN1000Inc4-8           41639         1029          -97.53%       
BenchmarkDscalIncN1000Inc10-8          46694         2014          -95.69%       
BenchmarkDscalIncN100000Inc1-8         100747        105440        +4.66%        
BenchmarkDscalIncN100000Inc2-8         131330        994430        +657.20%      
BenchmarkDscalIncN100000Inc4-8         157304        149662        -4.86%        
BenchmarkDscalIncN100000Inc10-8        520582        418432        -19.62%       
BenchmarkDscalIncN100000IncM1-8        113590        2063409       +1716.54%     
BenchmarkDscalIncN100000IncM2-8        4286527       113596        -97.35%       
BenchmarkDscalIncN100000IncM4-8        3628200       916309        -74.74%       
BenchmarkDscalIncN100000IncM10-8       526780        451271        -14.33%       
BenchmarkDscalIncToN1Inc1-8            6.43          5.90          -8.24%        
BenchmarkDscalIncToN2Inc1-8            7.54          7.07          -6.23%        
BenchmarkDscalIncToN2Inc2-8            7.44          7.07          -4.97%        
BenchmarkDscalIncToN2Inc4-8            7.55          6.79          -10.07%       
BenchmarkDscalIncToN2Inc10-8           7.70          7.19          -6.62%        
BenchmarkDscalIncToN3Inc1-8            9.14          7.59          -16.96%       
BenchmarkDscalIncToN3Inc2-8            9.11          7.65          -16.03%       
BenchmarkDscalIncToN3Inc4-8            8.93          7.37          -17.47%       
BenchmarkDscalIncToN3Inc10-8           8.90          7.16          -19.55%       
BenchmarkDscalIncToN4Inc1-8            11.5          8.85          -23.04%       
BenchmarkDscalIncToN4Inc2-8            11.2          8.97          -19.91%       
BenchmarkDscalIncToN4Inc4-8            11.5          9.23          -19.74%       
BenchmarkDscalIncToN4Inc10-8           11.4          8.80          -22.81%       
BenchmarkDscalIncToN10Inc1-8           16.9          13.1          -22.49%       
BenchmarkDscalIncToN10Inc2-8           16.7          12.9          -22.75%       
BenchmarkDscalIncToN10Inc4-8           17.0          13.0          -23.53%       
BenchmarkDscalIncToN10Inc10-8          16.9          12.8          -24.26%       
BenchmarkDscalIncToN1000Inc1-8         1038          1069          +2.99%        
BenchmarkDscalIncToN1000Inc2-8         1030          1112          +7.96%        
BenchmarkDscalIncToN1000Inc4-8         1311          1335          +1.83%        
BenchmarkDscalIncToN1000Inc10-8        2630          2594          -1.37%        
BenchmarkDscalIncToN100000Inc1-8       103488        106343        +2.76%        
BenchmarkDscalIncToN100000Inc2-8       134434        120745        -10.18%       
BenchmarkDscalIncToN100000Inc4-8       329089        296390        -9.94%        
BenchmarkDscalIncToN100000Inc10-8      1083672       1073600       -0.93%        
BenchmarkDscalIncToN100000IncM1-8      101419        108235        +6.72%        
BenchmarkDscalIncToN100000IncM2-8      135347        120553        -10.93%       
BenchmarkDscalIncToN100000IncM4-8      302446        297951        -1.49%        
BenchmarkDscalIncToN100000IncM10-8     1115914       1115239       -0.06% 
````

I also see the random spikes, but they do not align with the spikes you see.