@vladimir-ch 2016-01-14
A final comment for completeness. I tried to modify the benchmarks so that the `x` and `dst` slices are allocated inside the for loop hoping that it would reduce the probability that `x` is in the cache when Dscal is called, something along:
```go
func benchmarkDscalUnitary(b *testing.B, n int) {
	b.StopTimer()
	tmp := randomSlice(n, 1)
	a := rand.Float64()
	for i := 0; i < b.N; i++ {
		x := make([]float64, len(tmp))
		copy(x, tmp)
		b.StartTimer()
		DscalUnitary(a, x)
		b.StopTimer()
		gs = x
	}
}
```
Such benchmarks take much longer to finish (13 hours to get the output below) but give more consistent results:
```
% benchstat dscal_newbench_noasm.txt dscal_newbench.txt
name                     old time/op  new time/op   delta
DscalUnitaryN10          36.2ns ± 0%   33.2ns ± 1%   -8.30%  (p=0.000 n=5+6)
DscalUnitaryN100          106ns ± 0%     53ns ± 0%  -49.57%  (p=0.008 n=5+5)
DscalUnitaryN1000         768ns ± 0%    247ns ± 0%  -67.77%  (p=0.008 n=5+5)
DscalUnitaryN10000       7.27µs ± 0%   2.52µs ± 0%  -65.42%  (p=0.002 n=6+6)
DscalUnitaryN100000      71.1µs ± 1%   38.9µs ± 0%  -45.36%  (p=0.004 n=6+5)
DscalUnitaryToN10        36.7ns ± 2%   33.8ns ± 1%   -8.08%  (p=0.002 n=6+6)
DscalUnitaryToN100        111ns ± 5%     55ns ± 0%  -50.38%  (p=0.002 n=6+6)
DscalUnitaryToN1000       878ns ± 1%    278ns ± 1%  -68.32%  (p=0.002 n=6+6)
DscalUnitaryToN10000     7.82µs ± 2%   3.65µs ± 0%  -53.35%  (p=0.002 n=6+6)
DscalUnitaryToN100000    84.9µs ± 1%   61.8µs ± 0%  -27.21%  (p=0.004 n=6+5)
DscalUnitaryToXN10       36.0ns ± 1%   32.6ns ± 1%   -9.35%  (p=0.002 n=6+6)
DscalUnitaryToXN100       113ns ± 1%     55ns ± 0%  -51.49%  (p=0.000 n=6+5)
DscalUnitaryToXN1000      849ns ± 1%    249ns ± 0%  -70.73%  (p=0.004 n=6+5)
DscalUnitaryToXN10000    8.17µs ± 2%   2.47µs ± 0%  -69.79%  (p=0.002 n=6+6)
DscalUnitaryToXN100000   72.1µs ± 8%   39.1µs ± 0%  -45.72%  (p=0.002 n=6+6)
DscalIncN10Inc1          38.3ns ± 1%   41.2ns ± 0%   +7.39%  (p=0.002 n=6+6)
DscalIncN10Inc2          38.2ns ± 1%   41.6ns ± 0%   +9.04%  (p=0.002 n=6+6)
DscalIncN10Inc4          38.4ns ± 0%   41.6ns ± 0%   +8.23%  (p=0.004 n=6+5)
DscalIncN10Inc10         39.4ns ± 1%   42.0ns ± 0%   +6.55%  (p=0.010 n=6+4)
DscalIncN1000Inc1         884ns ± 1%   1038ns ± 0%  +17.44%  (p=0.004 n=5+6)
DscalIncN1000Inc2         895ns ± 0%   1040ns ± 0%  +16.22%  (p=0.004 n=5+6)
DscalIncN1000Inc4         937ns ± 1%   1053ns ± 0%  +12.40%  (p=0.002 n=6+6)
DscalIncN1000Inc10       1.65µs ± 0%   1.63µs ± 0%   -1.69%  (p=0.002 n=6+6)
DscalIncN100000Inc1      82.1µs ± 1%  100.7µs ± 0%  +22.70%  (p=0.002 n=6+6)
DscalIncN100000Inc2      99.2µs ± 0%  104.0µs ± 0%   +4.86%  (p=0.002 n=6+6)
DscalIncN100000Inc4       168µs ± 0%    163µs ± 0%   -3.03%  (p=0.008 n=5+5)
DscalIncN100000Inc10      456µs ± 0%    516µs ± 1%  +13.17%  (p=0.002 n=6+6)
DscalIncN100000IncM1     81.6µs ± 2%  100.7µs ± 0%  +23.44%  (p=0.002 n=6+6)
DscalIncN100000IncM2      103µs ± 0%    105µs ± 0%   +2.08%  (p=0.004 n=6+5)
DscalIncN100000IncM4      167µs ± 0%    161µs ± 0%   -3.17%  (p=0.002 n=6+6)
DscalIncN100000IncM10     414µs ± 0%    451µs ± 0%   +8.98%  (p=0.002 n=6+6)
DscalIncToN10Inc1        41.5ns ± 0%   43.3ns ± 1%   +4.34%  (p=0.002 n=6+6)
DscalIncToN10Inc2        41.8ns ± 0%   43.4ns ± 1%   +3.79%  (p=0.002 n=6+6)
DscalIncToN10Inc4        42.1ns ± 1%   43.4ns ± 1%   +3.25%  (p=0.002 n=6+6)
DscalIncToN10Inc10       41.7ns ± 0%   43.3ns ± 0%   +3.74%  (p=0.004 n=5+6)
DscalIncToN1000Inc1       903ns ± 0%   1041ns ± 0%  +15.29%  (p=0.002 n=6+6)
DscalIncToN1000Inc2       928ns ± 0%   1044ns ± 0%  +12.52%  (p=0.004 n=6+5)
DscalIncToN1000Inc4      1.21µs ± 0%   1.19µs ± 0%   -1.24%  (p=0.004 n=5+6)
DscalIncToN1000Inc10     3.00µs ± 0%   2.93µs ± 0%   -2.43%  (p=0.002 n=6+6)
DscalIncToN100000Inc1    92.1µs ± 0%  100.8µs ± 0%   +9.53%  (p=0.004 n=5+6)
DscalIncToN100000Inc2     141µs ± 0%    134µs ± 0%   -4.47%  (p=0.002 n=6+6)
DscalIncToN100000Inc4     293µs ± 1%    286µs ± 0%   -2.31%  (p=0.002 n=6+6)
DscalIncToN100000Inc10   1.07ms ± 1%   1.07ms ± 1%     ~     (p=0.310 n=6+6)
DscalIncToN100000IncM1   92.7µs ± 0%  100.8µs ± 0%   +8.78%  (p=0.004 n=6+5)
DscalIncToN100000IncM2    141µs ± 0%    133µs ± 0%   -5.73%  (p=0.002 n=6+6)
DscalIncToN100000IncM4    288µs ± 0%    281µs ± 0%   -2.28%  (p=0.008 n=5+5)
DscalIncToN100000IncM10   957µs ± 1%    971µs ± 0%   +1.40%  (p=0.002 n=6+6)
```
The times for `Inc` versions are still a bit strange but I'm leaving it for future work.